Test for length of output from entrypoint.sh
Discovered a case that a shell script without a shebang line will fail
to output anything, and will also fail to report an error to exec.Cmd's
caller.

This patch extends the logwriter wrapper to count the number of bytes
written and emit an warning on the logger if the entrypoint.sh logs
nothing. It is implemented as a warning on the logger rather than a
runtime error because it is forseeable that an entrypoint script may do
nothing, and we prefer warning instead of info, as this is very possibly
an error condition, not an info.

The tests are extended with an example log spy which tracks its output
and asserts accordingly.